refactor: 🔨 페이지네이션 공통 컴포넌트로 수정
