+++ src/search/PostListContainer.tsx
@@ -1,7 +1,8 @@
 import PostList from './PostList';
-import Pagination from './Pagination';
+
 import { useState, useEffect } from 'react';
 import { IsJsonString } from './isJsonString';
+import { Pagination } from '../common';
 
 interface Ilikes {
   _id: string;
@@ -106,7 +107,7 @@ const PostListContainer = ({
       <Pagination
         limit={limit}
         page={page}
-        totalPosts={filterPosts().length}
+        total={filterPosts().length}
         setPage={setPage}
         currentChannelId={currentChannelId}
       />
